Made neural networks with arrays
This allows significantly better and faster learning and computations, as well as plainly more comforatable

+++ Four in a row/Network/Network.cs
@@ -45,7 +45,7 @@ namespace Network
             }
         }
 
-        public void Feed(int[] input)
+        public void Feed(double[] input)
         {
             if (input.Length == Layers[0].Neurons.Count)
             {
